confluent/containers/app/fastapi/requirements.txt
asyncpg==0.30.0
fastapi==0.95.2
SQLAlchemy==1.4.54
psycopg2-binary==2.9.10
uvicorn==0.23.2
mlflow==2.22.4
pandas==1.5.3
pydantic==1.10.21
tensorflow==2.13.1
Pillow==12.1.1
python-multipart==0.0.22
fsspec==2024.10.0
s3fs==2024.10.0
boto3==1.37.0
scikit-learn==1.3.2
